BUG/MINOR: http-act: Fix parsing of the expression argument for pause action
When the "pause" action is parsed, if an expression is used instead of a
static value, the position of the current argument after the expression
evaluation is incremented while it should not. The sample_parse_expr()
function already take care of it. However, it should still be incremented
when an time value was parsed.

This patch must be backported to 3.2.

(cherry picked from commit 5da4da0bb62b6492c3a6fc31bd699220491bf21a)
Signed-off-by: William Lallemand <wlallemand@haproxy.com>