BUILD: quic: QUIC build against OpenSSL 3.5 broken
This commit was backported after the QUIC backend implementation.
This latter removes the ->li quic_conn struct member with this commit

   MINOR: quic-be: get rid of ->li quic_conn member

which has not to be backported.

That said, at this time 3.2 does not support the QUIC backends. So there
is no need to inspect ->li quic_conn struct member from the impacted function.

Simply replace the <!l> variable value by 0, <l> being always not NULL.

Thank you to @staticglobal for having reported this issue in GH #3027