BUG/MEDIUM: acme: use POST-as-GET instead of GET for resources
The requests that checked the status of the challenge and the retrieval
of the certificate were done using a GET.

This is working with letsencrypt and other CA providers, but it might
not work everywhere. RFC 8555 specifies that only the directory and
newNonce resources MUST work with a GET requests, but everything else
must use POST-as-GET.

Must be backported to 3.2.

(cherry picked from commit 7139ebd67659748e4447613cb3f2e12fd0229c12)
Signed-off-by: William Lallemand <wlallemand@haproxy.com>